chore: (docs) update tabs component demo (#1703)
* chore: (docs) update tabs component demo
